+++ packages/livechat/src/helpers/canRenderMessage.ts
@@ -25,6 +25,3 @@ const msgTypesNotRendered = [
 ];
 
 export const canRenderMessage = ({ t }: { t: string }) => !msgTypesNotRendered.includes(t);
-
-export const canRenderTriggerMessage = (user: { token: string }) => (message: { trigger?: boolean; triggerAfterRegistration?: boolean }) =>
-	!message.trigger || (!user && !message.triggerAfterRegistration) || (user && message.triggerAfterRegistration);